lambda/requirements.txt
python-docx==1.1.0
PyMuPDF==1.23.22
pytesseract==0.3.10
Pillow==10.3.0
pymongo==4.7.2
python-dotenv==1.0.1
pandas==2.2.2
python-pptx==0.6.23
beautifulsoup4==4.12.3
requests

